Refactor: split main components
