+++ src/render/draw.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include "../camera/camera.hpp"
+
+void drawGround();
+void drawTracks();
+void drawGrid();
+bool pickTile(const IsoCam& c, int& gx, int& gy);
