+++ src/render/models.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include "raylib.h"
+
+extern Model gEW, gNS, gDG;
+
+void initModels();
